services/inventory-backend/requirements/dev.txt
# ======================================================================
# DEVELOPMENT REQUIREMENTS
# Local development and tooling dependencies. These packages assist with
# interactive debugging, testing, linting, formatting, and development
# workflows. None of these should be installed in production.
# ======================================================================

-r base.txt

# Interactive shell utilities
ipython>=8.25,<9.0

# Code quality and formatting
black>=24.8,<25.0
ruff>=0.5,<1.0
isort>=5.13,<6.0

# Type checking
mypy>=1.11,<2.0
django-stubs>=5.0,<6.0
django-stubs-ext>=5.0,<6.0
types-python-dateutil>=2.9,<3.0

# Testing framework and related tools
pytest>=8.2,<9.0
pytest-django>=4.8,<5.0
pytest-cov>=5.0,<6.0

# Dependency security
pip-audit>=2.7,<3.0
